requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
ta>=0.11.0
vectorbt>=0.26.0
scikit-learn>=1.3.0
optuna>=3.0.0
PyPortfolioOpt>=1.5.0
PyYAML>=6.0
SQLAlchemy>=2.0.0
